메인화면, 연차 갯수, 프로젝트 합치기
수정사항
헤버 본인계정 클릭시 내 게시글 보기쪽으로 넘어가기
메인 페이지로 넘어가는 아이콘들 정리
근태 관련 코드 수정
메인 이동 페이지들 전부 수정(메인페이지 에서 메일함 게시글 근태등 이동 하기)
12.16 김우석

+++ devwareProject/src/main/java/com/oracle/devwareProject/dto/KiWoSu/Vacation.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class Vacation {
 	private int emp_num;
-	private int va_stock;
+	private float va_stock;
 }
